Rename bps -> bpsraw, bpsdmplex -> bps, areadmplex -> area
