+++ benchmarks/README.md
@@ -7,7 +7,7 @@ backends.
 
 Example:
 ```sh
-benchmark.sh -c /cpu/self -r petsc-bps.sh -b bp1 -n 16 -p 16
+benchmark.sh -c /cpu/self -r petsc-bpsraw.sh -b bp1 -n 16 -p 16
 ```
 where the option `-c <specs-list>` specifies a list of libCEED specs to
 benchmark, `-b <bp-list>` specifies a list of CEED benchmark problems to run,
@@ -17,14 +17,14 @@ of processors per node.
 Multiple backends, benchmark problems, and processor configurations can be
 benchmarked with:
 ```sh
-benchmark.sh -c "/cpu/self/ref/serial /cpu/self/ref/blocked" -r petsc-bps.sh -b "bp1 bp3" -n "16 32 64" -p "16 32 64"
+benchmark.sh -c "/cpu/self/ref/serial /cpu/self/ref/blocked" -r petsc-bpsraw.sh -b "bp1 bp3" -n "16 32 64" -p "16 32 64"
 ```
 
 The results from the benchmarks are written to files named `*-output.txt`.
 
 For a short help message, use the option `-h`.
 
-When running the tests `petsc-bps.sh`, the following
+When running the tests `petsc-bpsraw.sh`, the following
 variables can be set on the command line:
 * `max_dofs_node=<number>`, e.g. `max_dofs_node=1000000` - this sets the upper
   bound of the problem sizes, per compute node; the default value is 3*2^20.
@@ -37,7 +37,7 @@ After generating the results, use the `postprocess-plot.py` script (which
 requires the python package matplotlib) or the `postprocess-table.py` script,
 e.g.:
 ```sh
-python postprocess-plot.py petsc-bps-bp1-*-output.txt
+python postprocess-plot.py petsc-bpsraw-bp1-*-output.txt
 ```
 The plot ranges and some other options can be adjusted by editing the values
 in the beginning of the script `postprocess-plot.py`.
